Application scaffold refactoring
- removed index.vue file
- added index.js and index.html as main application layout
- reduced comments in bootstrap.js file, which makes it more readable, and moved all imports on the top
- moved helpers outside the app folder
- added lib folder in src/

Store refactoring
- folder name is now changed to vuex and main file is store.js
- added module imports instead of single actions and getters
- added index.js in each module
