--- src/app/index.vue
@@ -1,31 +1,0 @@
-<template>
-  <div>
-    <router-view></router-view>
-  </div>
-</template>
-<script>
-  /* ============
-   * Entry Point
-   * ============
-   *
-   * The entry point of the application
-   */
-  import store from './store'
-  import { router } from './../bootstrap'
-
-  export default {
-    /**
-     * The Vuex store
-     */
-    store,
-    /**
-     * The router
-     */
-    router,
-    /**
-     * Fires when the app has been mounted
-     */
-    mounted () {
-    }
-  }
-</script>
