+++ src/app/pages/home/home.html
@@ -1,12 +1,10 @@
-<layout>
-  <div class="content">
-    <div class="content-inner">
-      <h1 class="title">
-        Vueture
-      </h1>
-      <small class="subtitle">
-        {{ $t('message')}}
-      </small>
-    </div>
+<div class="content">
+  <div class="content-inner">
+    <h1 class="title">
+      Vueture
+    </h1>
+    <small class="subtitle">
+      {{ $t('message')}}
+    </small>
   </div>
-</layout>
+</div>
