+++ src/app/pages/home/home.js
@@ -1,9 +1,7 @@
 import template from './home.html'
-import layout from '../../layouts/default/default'
 
 require('./home.scss')
 
 export default {
-  components: { layout },
   template
 }
