+++ src/app/vuex/modules/locale/events.js
