+++ src/app/vuex/modules/locale/getters.js
@@ -1,0 +1,2 @@
+export const languages = state => state.languages
+export const activeLanguageCode = state => state.language.code
