+++ src/app/vuex/modules/locale/index.js
@@ -1,0 +1,10 @@
+import { state, mutations } from './mutations'
+import actions from './actions'
+import * as getters from './getters'
+
+export default {
+  state,
+  mutations,
+  actions,
+  getters
+}
