+++ src/bootstrap.js
@@ -1,84 +1,44 @@
-/* ============
- * Bootstrap File
- * ============
- *
- * Will configure and bootstrap the application
- */
-
-/* ============
- * Vue
- * ============
- *
- * Vue.js is a library for building interactive web interfaces.
- * It provides data-reactive components with a simple and flexible API.
- *
- * http://rc.vuejs.org/guide/
- */
 import Vue from 'vue'
+import VueResource from 'vue-resource'
+import { sync } from 'vuex-router-sync'
+import VueRouter from 'vue-router'
 
-Vue.config.debug = process.env.NODE_ENV !== 'production'
+import routes from './app/routes'
+import store from './app/vuex/store'
 
-/* ============
- * Vue Resource
- * ============
- *
- * Vue Resource provides services for making web requests and handle
- * responses using a XMLHttpRequest or JSONP.
- *
+/**
+ * Initialize the main sass file of the application style
+ */
+require('./assets/sass/app.scss')
+
+/**
+ * Initialize vue-resource plugin to manage http requests
  * https://github.com/vuejs/vue-resource/tree/master/docs
  */
-import VueResource from 'vue-resource'
-
 Vue.use(VueResource)
 
 Vue.http.headers.common.Accept = 'application/json'
 Vue.http.options.root = process.env.API_LOCATION
 
-/* ============
- * Vuex Router Sync
- * ============
- *
- * Effortlessly keep vue-Router and vuex store in sync.
- *
- * https://github.com/vuejs/vuex-router-sync/blob/master/README.md
- */
-import VuexRouterSync from 'vuex-router-sync'
-import store from './app/store'
-
-/* ============
- * Vue Router
- * ============
- *
- * The official Router for Vue.js. It deeply integrates with Vue.js core
- * to make building Single Page Applications with Vue.js a breeze.
- *
+/**
+ * Initialize vue-resource plugin to manage application routing
  * http://router.vuejs.org/en/index.html
  */
-import VueRouter from 'vue-router'
-import routes from './app/routes'
-
 Vue.use(VueRouter)
 
-export const router = new VueRouter({
+const router = new VueRouter({
   mode: 'history',
   routes
 })
 
-VuexRouterSync.sync(store, router)
-
-Vue.router = router
-
-/* ============
- * Styling
- * ============
- *
- * Vueture comes with SASS out of the box.
- * If you want to use a different CSS preprocessor,
- * install the desired Webpack loader.
- * Require the entrypoint here.
+/**
+ * Sync routes with vuex store
  */
-require('./assets/sass/app.scss')
+sync(store, router)
 
-export default {
-  router
-}
+/**
+ * The debug mode is available globally in the Vue.config.debug property
+ */
+Vue.config.debug = process.env.NODE_ENV !== 'production'
+
+export { router, store }
